+++ config/gravatar.php
@@ -77,10 +77,11 @@ return [
             | - null value to fallback to the default Gravatar
             | - a string represanting the URL of your own default image
             | - '404': do not load any image if none is associated with the email hash, instead return an HTTP 404 (File Not Found) response
-            | - 'mm': (mystery-man) a simple, cartoon-style silhouetted outline of a person (does not vary by email hash)
+            | - 'mm': (mystery-person) a simple, cartoon-style silhouetted outline of a person (does not vary by email hash)
             | - 'identicon': a geometric pattern based on an email hash
             | - 'monsterid': a generated 'monster' with different colors, faces, etc
             | - 'wavatar': generated faces with differing features and backgrounds
+            | - 'robohash': a generated robot with different colors, faces, etc
             | - 'retro': awesome generated, 8-bit arcade-style pixelated faces
             | - 'blank': a transparent PNG image
             */
@@ -145,11 +146,11 @@ return [
 
             'size' => 120,
 
-            'default_image' => 'mm',
+            'default_image' => 'mp',
 
             'force_default' => false,
 
-            'max_rating' => 'r',
+            'max_rating' => 'g',
 
             'extension' => 'jpg',