+++ src/ServiceProvider.php
@@ -1,24 +1,26 @@
 <?php
 
-namespace forxer\LaravelGravatar;
+namespace LaravelGravatar;
 
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
 
 class ServiceProvider extends BaseServiceProvider
 {
-    public function boot()
-    {
-        $this->publishes([
-            __DIR__ . '/../config/gravatar.php' => $this->app->configPath() . '/' . 'gravatar.php',
-        ], 'config');
-    }
-
     public function register()
     {
-        $this->mergeConfigFrom(__DIR__ . '/../config/gravatar.php', 'gravatar');
+        $this->mergeConfigFrom(__DIR__.'/../config/gravatar.php', 'gravatar');
 
         $this->app->singleton('gravatar', function ($app) {
-            return new Gravatar($app);
+            return new Gravatar($app['config']['gravatar']);
         });
+    }
+
+    public function boot()
+    {
+        if ($this->app->runningInConsole()) {
+            $this->publishes([
+                __DIR__ . '/../config/gravatar.php' => $this->app->configPath('gravatar.php'),
+            ], 'gravatar-config');
+        }
     }
 }